[refactor] 챗봇 관련 컴포넌트 파일 분리 - types/index.ts 파일 생성, Message interface 정의 #22
