+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export interface Message {
+  text: string;
+  isUser: boolean;
+  isTyping?: boolean;
+  id?: number;
+}
